feat: refactor configuration and model structure; soc new schemas and utility functions for tsunami calculations
